Reusing code for the Pathway Explorer
